+++ chromadb/api/fastapi.py
@@ -6,6 +6,7 @@ from uuid import UUID
 import requests
 from overrides import override
 
+from chromadb.api.base_http_client import BaseHTTPClient
 import chromadb.errors as errors
 from chromadb.types import Database, Tenant
 import chromadb.utils.embedding_functions as ef
@@ -39,63 +40,12 @@ from chromadb.telemetry.opentelemetry import (
     trace_method,
 )
 from chromadb.telemetry.product import ProductTelemetryClient
-from urllib.parse import urlparse, urlunparse, quote
 from chromadb.types import Collection as CollectionModel
 
 logger = logging.getLogger(__name__)
 
 
-class FastAPI(ServerAPI):
-    _settings: Settings
-    _max_batch_size: int = -1
-
-    @staticmethod
-    def _validate_host(host: str) -> None:
-        parsed = urlparse(host)
-        if "/" in host and parsed.scheme not in {"http", "https"}:
-            raise ValueError(
-                "Invalid URL. " f"Unrecognized protocol - {parsed.scheme}."
-            )
-        if "/" in host and (not host.startswith("http")):
-            raise ValueError(
-                "Invalid URL. "
-                "Seems that you are trying to pass URL as a host but without \
-                    specifying the protocol. "
-                "Please add http:// or https:// to the host."
-            )
-
-    @staticmethod
-    def resolve_url(
-        chroma_server_host: str,
-        chroma_server_ssl_enabled: Optional[bool] = False,
-        default_api_path: Optional[str] = "",
-        chroma_server_http_port: Optional[int] = 8000,
-    ) -> str:
-        _skip_port = False
-        _chroma_server_host = chroma_server_host
-        FastAPI._validate_host(_chroma_server_host)
-        if _chroma_server_host.startswith("http"):
-            logger.debug("Skipping port as the user is passing a full URL")
-            _skip_port = True
-        parsed = urlparse(_chroma_server_host)
-
-        scheme = "https" if chroma_server_ssl_enabled else parsed.scheme or "http"
-        net_loc = parsed.netloc or parsed.hostname or chroma_server_host
-        port = (
-            ":" + str(parsed.port or chroma_server_http_port) if not _skip_port else ""
-        )
-        path = parsed.path or default_api_path
-
-        if not path or path == net_loc:
-            path = default_api_path if default_api_path else ""
-        if not path.endswith(default_api_path or ""):
-            path = path + default_api_path if default_api_path else ""
-        full_url = urlunparse(
-            (scheme, f"{net_loc}{port}", quote(path.replace("//", "/")), "", "", "")
-        )
-
-        return full_url
-
+class FastAPI(BaseHTTPClient, ServerAPI):
     def __init__(self, system: System):
         super().__init__(system)
         system.settings.require("chroma_server_host")
